requirements.txt
asgiref==3.11.0
Django==6.0.1
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
model-bakery==1.23.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
sqlparse==0.5.5
uuid==1.30
